za6_tools:  Fix tool top-level xacro for build & display, both
When running the `display_*.launch` files, `$(cwd)` doesn't work,
because the current directory is the `urdf/` directory.  When building
the static URDF files from `CMakeLists.txt`, `$(find za_tools)`
doesn't work, because it refers to itself, which is currently being
built.

Fix both cases by hacking in a `_build` arg to select the appropriate
method.