za6_tools:  Fix gripper URDF xacros for MoveIt Studio
Static URDF used as Studio `web_ui_gripper` was broken.  The iMarker
appeared at the robot base and couldn't be used to plan moves.

The meaningful change here is removing the joint that connected the
`grasp_link` link, which doesn't exist.

The versabuilt_multigrip is broken.  The scale looks off, it seems to
be missing `za6_multigrip_adapter` geometry incl. meshes, and
orientation is wrong.